windows portable: dev CI tests
