+++ .github/workflows/analysis-coverage.yml
@@ -19,12 +19,12 @@ jobs:
 
     steps:
       - uses: actions/checkout@v4
-#      - uses: actions/setup-python@v5
-#        with:
-#          python-version: "3.10"
-#
-#      - name: Install from source
-#        run: python3 -m pip install ".[dev]"
-#
-#      - name: Run Analysis
-#        run: python3 -m pylint "visionatrix/"
+      - uses: actions/setup-python@v5
+        with:
+          python-version: "3.10"
+
+      - name: Install from source
+        run: python3 -m pip install ".[dev]"
+
+      - name: Run Analysis
+        run: python3 -m pylint "visionatrix/"
