+++ .github/workflows/openapi-check.yml
@@ -15,21 +15,21 @@ jobs:
 
     steps:
       - uses: actions/checkout@v4
-#      - uses: actions/setup-python@v5
-#        with:
-#          python-version: "3.10"
-#
-#      - name: Install PyTorch for CPU
-#        run: python3 -m pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
-#
-#      - name: Install from source
-#        run: python3 -m pip install ".[dev]"
-#
-#      - name: Save current OpenAPI specs
-#        run: cp openapi.json original_openapi.json
-#
-#      - name: Generate new OpenAPI specs
-#        run: make openapi && pre-commit run --all-files || true
-#
-#      - name: Compare OpenAPI specs
-#        run: python3 scripts/ci/openapi_compare.py openapi.json original_openapi.json
+      - uses: actions/setup-python@v5
+        with:
+          python-version: "3.10"
+
+      - name: Install PyTorch for CPU
+        run: python3 -m pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
+
+      - name: Install from source
+        run: python3 -m pip install ".[dev]"
+
+      - name: Save current OpenAPI specs
+        run: cp openapi.json original_openapi.json
+
+      - name: Generate new OpenAPI specs
+        run: make openapi && pre-commit run --all-files || true
+
+      - name: Compare OpenAPI specs
+        run: python3 scripts/ci/openapi_compare.py openapi.json original_openapi.json
